Move notifications interfaces in UserNotifications class
