--- src/types/Notification.ts
@@ -1,15 +1,0 @@
-import ChargingStation from './ChargingStation';
-import User from './User';
-
-export interface NotifySessionNotStarted {
-  chargingStation: ChargingStation;
-  tagID: string;
-  authDate: Date;
-  user: User;
-}
-export interface AdminAccountVerificationNotification {
-  user: User;
-  evseDashboardURL: string;
-  evseUserToVerifyURL: string;
-  tenantLogoURL: string;
-}
